Bring HP system and more improvements to Shiken (#204)
* Remove info, id, gifid, stats from misc
* Added info, id, gifid, stats to userinfo
* Added missing import 1
* Minor correction to markdown help
* [user info] Added hp bar maker
* Fix typo
* Added hp bar
* Added newline and make text bold
* Userinfo: add hpmanager function
* hpmanager: smol fix
* fix mistake in logic
* Swap variables positions
* Pretty print telegram ID
* Update the title text in info
* Update no setme defined response
* Update to replies on responses
* Update to no setbio defined response
* More updates to replies on responses
* Monospace health and swap variables"
* Added hp bar

* Make percentage integer

* Edit info text to cover waiting time

* try bar 11
* Dont show gban/blakclist thingies for unbannable users
* Edit to appraisal text
* Use the 5th progress bar version
* Replace bar maker
* Make bar italics
* Add profile pic in userinfo
* Smol fix
* Edit appraisal text
* Edits to presence detect
* [misc] Fixed markdown typos as dictated in #198
* Fixed a mis-paste
* Edit the header text on info
* Wrap health around braces
* Remove bold from presence
* Trim chatcount response text
* Userinfo: make it reply to msg
