+++ SaitamaRobot/modules/misc.py
@@ -1,31 +1,24 @@
-import html
-import re
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup
-import requests
-from SaitamaRobot import (DEV_USERS, OWNER_ID, SUDO_USERS, SUPPORT_USERS,
-                          TIGER_USERS, WHITELIST_USERS, dispatcher, sw)
-from SaitamaRobot.__main__ import STATS, TOKEN, USER_INFO
+from SaitamaRobot.modules.helper_funcs.chat_status import user_admin
 from SaitamaRobot.modules.disable import DisableAbleCommandHandler
-from SaitamaRobot.modules.sql.afk_sql import is_afk
-from SaitamaRobot.modules.helper_funcs.chat_status import sudo_plus, user_admin
-from SaitamaRobot.modules.helper_funcs.extraction import extract_user
+from SaitamaRobot import dispatcher
+
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 from telegram import MessageEntity, ParseMode, Update
-from telegram.error import BadRequest
-from telegram.ext import CallbackContext, CommandHandler, Filters, run_async
-from telegram.utils.helpers import mention_html
+from telegram.ext.dispatcher import run_async
+from telegram.ext import CallbackContext, Filters, CommandHandler
 
 MARKDOWN_HELP = f"""
 Markdown is a very powerful formatting tool supported by telegram. {dispatcher.bot.first_name} has some enhancements, to make sure that \
 saved messages are correctly parsed, and to allow you to create buttons.
 
-• <code>_italic_</code>*:* wrapping text with '_' will produce italic text
-• <code>*bold*</code>*:* wrapping text with '*' will produce bold text
-• <code>`code`</code>*:* wrapping text with '`' will produce monospaced text, also known as 'code'
-• <code>[sometext](someURL)</code>*:* this will create a link - the message will just show <code>sometext</code>, \
+• <code>_italic_</code>: wrapping text with '_' will produce italic text
+• <code>*bold*</code>: wrapping text with '*' will produce bold text
+• <code>`code`</code>: wrapping text with '`' will produce monospaced text, also known as 'code'
+• <code>[sometext](someURL)</code>: this will create a link - the message will just show <code>sometext</code>, \
 and tapping on it will open the page at <code>someURL</code>.
-<b>Example:</b>Example:<b>Example:</b> <code>[test](example.com)</code>
+<b>Example:</b><code>[test](example.com)</code>
 
-• <code>[buttontext](buttonurl:someURL)</code>*:* this is a special enhancement to allow users to have telegram \
+• <code>[buttontext](buttonurl:someURL)</code>: this is a special enhancement to allow users to have telegram \
 buttons in their markdown. <code>buttontext</code> will be what is displayed on the button, and <code>someurl</code> \
 will be the url which is opened.
 <b>Example:</b> <code>[This is a button](buttonurl:example.com)</code>
@@ -37,174 +30,6 @@ This will create two buttons on a single line, instead of one button per line.
 
 Keep in mind that your message <b>MUST</b> contain some text other than just a button!
 """
-
-
-@run_async
-def get_id(update: Update, context: CallbackContext):
-    bot, args = context.bot, context.args
-    message = update.effective_message
-    chat = update.effective_chat
-    msg = update.effective_message
-    user_id = extract_user(msg, args)
-
-    if user_id:
-
-        if msg.reply_to_message and msg.reply_to_message.forward_from:
-
-            user1 = message.reply_to_message.from_user
-            user2 = message.reply_to_message.forward_from
-
-            msg.reply_text(
-                f"The original sender, {html.escape(user2.first_name)},"
-                f" has an ID of <code>{user2.id}</code>.\n"
-                f"The forwarder, {html.escape(user1.first_name)},"
-                f" has an ID of <code>{user1.id}</code>.",
-                parse_mode=ParseMode.HTML)
-
-        else:
-
-            user = bot.get_chat(user_id)
-            msg.reply_text(
-                f"{html.escape(user.first_name)}'s id is <code>{user.id}</code>.",
-                parse_mode=ParseMode.HTML)
-
-    else:
-
-        if chat.type == "private":
-            msg.reply_text(
-                f"Your id is <code>{chat.id}</code>.",
-                parse_mode=ParseMode.HTML)
-
-        else:
-            msg.reply_text(
-                f"This group's id is <code>{chat.id}</code>.",
-                parse_mode=ParseMode.HTML)
-
-
-@run_async
-def gifid(update: Update, context: CallbackContext):
-    msg = update.effective_message
-    if msg.reply_to_message and msg.reply_to_message.animation:
-        update.effective_message.reply_text(
-            f"Gif ID:\n<code>{msg.reply_to_message.animation.file_id}</code>",
-            parse_mode=ParseMode.HTML)
-    else:
-        update.effective_message.reply_text(
-            "Please reply to a gif to get its ID.")
-
-
-@run_async
-def info(update: Update, context: CallbackContext):
-    bot, args = context.bot, context.args
-    message = update.effective_message
-    chat = update.effective_chat
-    user_id = extract_user(update.effective_message, args)
-
-    if user_id:
-        user = bot.get_chat(user_id)
-
-    elif not message.reply_to_message and not args:
-        user = message.from_user
-
-    elif not message.reply_to_message and (
-            not args or
-        (len(args) >= 1 and not args[0].startswith("@") and
-         not args[0].isdigit() and
-         not message.parse_entities([MessageEntity.TEXT_MENTION]))):
-        message.reply_text("I can't extract a user from this.")
-        return
-
-    else:
-        return
-
-    text = (f"<b>Characteristics:</b>\n"
-            f"ID: <code>{user.id}</code>\n"
-            f"First Name: {html.escape(user.first_name)}")
-
-    if user.last_name:
-        text += f"\nLast Name: {html.escape(user.last_name)}"
-
-    if user.username:
-        text += f"\nUsername: @{html.escape(user.username)}"
-
-    text += f"\nPermalink: {mention_html(user.id, 'link')}"
-
-    if chat.type != "private":
-        _stext = "\nStatus: {}"
-
-        afk_st = is_afk(user.id)
-        if afk_st:
-            text += _stext.format("Sleeping")
-        else:
-            status = status = bot.get_chat_member(chat.id, user.id).status
-            if status:
-                if status in {"left", "kicked"}:
-                    text += _stext.format("Absent")
-                elif status == "member":
-                    text += _stext.format("Present")
-                elif status in {"administrator", "creator"}:
-                    text += _stext.format("Admin")
-
-    try:
-        spamwtc = sw.get_ban(int(user.id))
-        if spamwtc:
-            text += "\n\n<b>This person is Spamwatched!</b>"
-            text += f"\nReason: <pre>{spamwtc.reason}</pre>"
-            text += "\nAppeal at @SpamWatchSupport"
-        else:
-            pass
-    except:
-        pass  # don't crash if api is down somehow...
-
-    disaster_level_present = False
-
-    if user.id == OWNER_ID:
-        text += "\n\nThe Disaster level of this person is 'God'."
-        disaster_level_present = True
-    elif user.id in DEV_USERS:
-        text += "\n\nThis member is one of 'Hero Association'."
-        disaster_level_present = True
-    elif user.id in SUDO_USERS:
-        text += "\n\nThe Disaster level of this person is 'Dragon'."
-        disaster_level_present = True
-    elif user.id in SUPPORT_USERS:
-        text += "\n\nThe Disaster level of this person is 'Demon'."
-        disaster_level_present = True
-    elif user.id in TIGER_USERS:
-        text += "\n\nThe Disaster level of this person is 'Tiger'."
-        disaster_level_present = True
-    elif user.id in WHITELIST_USERS:
-        text += "\n\nThe Disaster level of this person is 'Wolf'."
-        disaster_level_present = True
-
-    if disaster_level_present:
-        text += ' [<a href="https://t.me/{}?start=disasters">?</a>]'.format(
-            bot.username)
-
-    try:
-        user_member = chat.get_member(user.id)
-        if user_member.status == 'administrator':
-            result = requests.post(
-                f"https://api.telegram.org/bot{TOKEN}/getChatMember?chat_id={chat.id}&user_id={user.id}"
-            )
-            result = result.json()["result"]
-            if "custom_title" in result.keys():
-                custom_title = result['custom_title']
-                text += f"\n\nThis user holds the title <b>{custom_title}</b> here."
-    except BadRequest:
-        pass
-
-    for mod in USER_INFO:
-        try:
-            mod_info = mod.__user_info__(user.id).strip()
-        except TypeError:
-            mod_info = mod.__user_info__(user.id, chat.id).strip()
-        if mod_info:
-            text += "\n\n" + mod_info
-
-    update.effective_message.reply_text(
-        text, parse_mode=ParseMode.HTML, disable_web_page_preview=True)
-
 
 @run_async
 @user_admin
@@ -228,7 +53,7 @@ def markdown_help_sender(update: Update):
     update.effective_message.reply_text(
         MARKDOWN_HELP, parse_mode=ParseMode.HTML)
     update.effective_message.reply_text(
-        "Try forwarding the following message to me, and you'll see!")
+        "Try forwarding the following message to me, and you'll see, and Use #test!")
     update.effective_message.reply_text(
         "/save test This is a markdown test. _italics_, *bold*, code, "
         "[URL](example.com) [button](buttonurl:github.com) "
@@ -247,14 +72,6 @@ def markdown_help(update: Update, context: CallbackContext):
             ]]))
         return
     markdown_help_sender(update)
-
-
-@run_async
-@sudo_plus
-def stats(update: Update, context: CallbackContext):
-    stats = "<b>📟 Current stats:</b>\n" + "\n".join([mod.__stats__() for mod in STATS])
-    result = re.sub(r'(\d+)', r'<code>\1</code>', stats)
-    update.effective_message.reply_text(result, parse_mode=ParseMode.HTML)
 
 
 __help__ = """
@@ -278,23 +95,17 @@ Example:
 Output: `1.0 USD = 75.505 INR`
 """
 
-ID_HANDLER = DisableAbleCommandHandler("id", get_id)
-GIFID_HANDLER = DisableAbleCommandHandler("gifid", gifid)
-INFO_HANDLER = DisableAbleCommandHandler(["info"], info)
-ECHO_HANDLER = DisableAbleCommandHandler("echo", echo, filters=Filters.group)
-MD_HELP_HANDLER = CommandHandler("markdownhelp", markdown_help)
-STATS_HANDLER = CommandHandler("stats", stats)
 
-dispatcher.add_handler(ID_HANDLER)
-dispatcher.add_handler(GIFID_HANDLER)
-dispatcher.add_handler(INFO_HANDLER)
+ECHO_HANDLER = DisableAbleCommandHandler("echo", echo, filters=Filters.group)
+MD_HELP_HANDLER = CommandHandler(
+    "markdownhelp", markdown_help)
+
 dispatcher.add_handler(ECHO_HANDLER)
 dispatcher.add_handler(MD_HELP_HANDLER)
-dispatcher.add_handler(STATS_HANDLER)
+
 
 __mod_name__ = "Extras"
-__command_list__ = ["id", "info", "echo"]
+__command_list__ = ["id", "echo"]
 __handlers__ = [
-    ID_HANDLER, GIFID_HANDLER, INFO_HANDLER, ECHO_HANDLER, MD_HELP_HANDLER,
-    STATS_HANDLER
+    ECHO_HANDLER, MD_HELP_HANDLER,
 ]
