+++ SaitamaRobot/modules/users.py
@@ -145,7 +145,7 @@ def __user_info__(user_id):
     if user_id == dispatcher.bot.id:
         return """I've seen them in... Wow. Are they stalking me? They're in all the same places I am... oh. It's me."""
     num_chats = sql.get_user_num_chats(user_id)
-    return f"""I've seen them in <code>{num_chats}</code> chats in total."""
+    return f"""I've seen them in <code>{num_chats}</code> groups."""
 
 
 def __stats__():
